Changed type of `hashtag_num` and `leading_quotes` to int.

+++ common/string_helpers.cpp
@@ -27,7 +27,7 @@ static auto FromHex(char c) -> std::optional<char> {
   return std::nullopt;
 }
 
-auto UnescapeStringLiteral(llvm::StringRef source, const size_t hashtag_num,
+auto UnescapeStringLiteral(llvm::StringRef source, const int hashtag_num,
                            bool is_block_string) -> std::optional<std::string> {
   std::string ret;
   ret.reserve(source.size());
@@ -105,7 +105,7 @@ auto UnescapeStringLiteral(llvm::StringRef source, const size_t hashtag_num,
   return ret;
 }
 
-auto ParseBlockStringLiteral(llvm::StringRef source, const size_t hashtag_num)
+auto ParseBlockStringLiteral(llvm::StringRef source, const int hashtag_num)
     -> ErrorOr<std::string> {
   llvm::SmallVector<llvm::StringRef> lines;
   source.split(lines, '\n', /*MaxSplit=*/-1, /*KeepEmpty=*/true);
